# modules/engine/src/main/java/uy/edu/utec/iiss/engine/rest/dto/ControladorEstadoResponse.java
package uy.edu.utec.iiss.engine.rest.dto;

/** Cuerpo de respuesta de los comandos /controlador/*. */
public class ControladorEstadoResponse {

    private boolean corriendo;

    public ControladorEstadoResponse() {
    }

    public ControladorEstadoResponse(boolean corriendo) {
        this.corriendo = corriendo;
    }

    public boolean isCorriendo() {
        return corriendo;
    }

    public void setCorriendo(boolean corriendo) {
        this.corriendo = corriendo;
    }
}